Rework server loading
